refact: bento lib decoupling
