+++ tests/test_experiments.py
@@ -9,6 +9,8 @@ from transcriptomics_data_service.models import ExperimentResult
 config = get_config()
 logger = get_logger(config)
 
+APPLICATION_JSON = {"content-type": "application/json"}
+
 TEST_EXPERIMENT_RESULT_NO_EXTRA = ExperimentResult(
     experiment_result_id="12345",
     assembly_id="assembly_test_id",
@@ -23,15 +25,8 @@ def test_create_experiment(exp, test_client, authz_headers, db_cleanup):
         headers=authz_headers,
         data=exp.model_dump_json(),
     )
+    logger.debug(response)
     assert response.status_code == status.HTTP_200_OK
-
-
-def test_create_experiment_400(test_client, authz_headers, db_cleanup):
-    response = test_client.post(
-        "/experiment",
-        headers=authz_headers,
-    )
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
 
 
 def test_create_experiment_403(test_client, authz_headers_bad, db_cleanup):
@@ -54,26 +49,11 @@ def test_get_experiment(test_client, authz_headers, db_with_experiment, db_clean
     assert TEST_EXPERIMENT_RESULT == ExperimentResult(**data)
 
 
-# EMPTY DB
-def test_get_experiment_400(test_client, db_with_experiment, db_cleanup):
-    # Missing api-key
-    response = test_client.get("/experiment/non-existant")
-    print(response)
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-
-
 def test_get_experiment_200_empty(test_client, authz_headers):
     # Missing api-key
     response = test_client.get("/experiment/non-existant", headers=authz_headers)
     assert response.status_code == status.HTTP_200_OK
     assert response.json() is None
-
-
-def test_delete_experiment_400(test_client, db_with_experiment, db_cleanup):
-    # Missing api-key
-    response = test_client.delete("/experiment/non-existant")
-    print(response)
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
 
 
 def test_delete_experiment_403(test_client, authz_headers_bad):
