+++ transcriptomics_data_service/constants.py
@@ -11,6 +11,8 @@ __all__ = [
 
 BENTO_SERVICE_KIND = "TDS"
 
+GENE_ID_KEY = "GeneID"
+
 SERVICE_GROUP = SERVICE_GROUP_BENTO
 SERVICE_ARTIFACT = BENTO_SERVICE_KIND
